# apps/api/openapi.json
{
  "components": {
    "schemas": {
      "AnalysisRunResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "profile_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Version"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/AnalysisStatus"
          }
        },
        "required": [
          "id",
          "project_id",
          "status",
          "created_at",
          "finished_at",
          "error",
          "profile_version"
        ],
        "title": "AnalysisRunResponse",
        "type": "object"
      },
      "AnalysisStatus": {
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed"
        ],
        "title": "AnalysisStatus",
        "type": "string"
      },
      "Audience": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "default": 0.0,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "title": "Evidence",
            "type": "array"
          },
          "pains": {
            "items": {
              "type": "string"
            },
            "title": "Pains",
            "type": "array"
          },
          "primary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary"
          }
        },
        "title": "Audience",
        "type": "object"
      },
      "Brand": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "default": 0.0,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "do_not_mention": {
            "items": {
              "type": "string"
            },
            "title": "Do Not Mention",
            "type": "array"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "title": "Evidence",
            "type": "array"
          },
          "palette": {
            "items": {
              "pattern": "^#[0-9A-Fa-f]{6}$",
              "type": "string"
            },
            "title": "Palette",
            "type": "array"
          },
          "voice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice"
          }
        },
        "title": "Brand",
        "type": "object"
      },
      "BusinessModel": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "default": 0.0,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "title": "Evidence",
            "type": "array"
          },
          "trial_days": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trial Days"
          },
          "type": {
            "$ref": "#/components/schemas/BusinessModelType",
            "default": "unknown"
          }
        },
        "title": "BusinessModel",
        "type": "object"
      },
      "BusinessModelType": {
        "enum": [
          "free",
          "subscription",
          "one_time_purchase",
          "ads",
          "unknown"
        ],
        "title": "BusinessModelType",
        "type": "string"
      },
...
      "ErrorCode": {
        "enum": [
          "invalid_repository_url",
          "not_local_request",
          "project_not_found",
          "analysis_not_found",
          "profile_not_found",
          "project_already_exists",
          "analysis_already_running",
          "profile_already_approved",
          "internal_error",
          "database_unavailable"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "Evidence": {
        "additionalProperties": false,
        "description": "A location in the analysed repository that supports a claim.",
        "properties": {
          "file": {
            "title": "File",
            "type": "string"
          },
          "lines": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lines"
          }
        },
        "required": [
          "file"
        ],
        "title": "Evidence",
        "type": "object"
      },
      "Feature": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "default": 0.0,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "description": {
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "title": "Evidence",
            "type": "array"
          },
          "id": {
            "maxLength": 64,
            "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$",
            "title": "Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/FeatureStatus",
            "default": "unknown"
          }
        },
        "required": [
          "id",
          "description"
        ],
        "title": "Feature",
        "type": "object"
      },
      "FeatureStatus": {
        "description": "The publishable flag: only ``live`` features may be mentioned publicly.",
        "enum": [
          "live",
          "unreleased",
          "unknown"
        ],
        "title": "FeatureStatus",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Measurement": {
        "additionalProperties": false,
        "properties": {
          "analytics": {
            "items": {
              "type": "string"
            },
            "title": "Analytics",
            "type": "array"
          },
          "attribution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attribution"
          },
          "confidence": {
            "default": 0.0,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "deep_links": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deep Links"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "title": "Evidence",
            "type": "array"
          }
        },
        "title": "Measurement",
        "type": "object"
      },
      "Problem": {
        "additionalProperties": false,
        "description": "The body of every error response except request validation (422).",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "Problem",
        "type": "object"
      },
      "Product": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "default": 0.0,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "title": "Evidence",
            "type": "array"
          },
          "languages": {
            "items": {
              "pattern": "^[a-z]{2,3}(-[A-Za-z0-9]{2,8})*$",
              "type": "string"
            },
            "title": "Languages",
            "type": "array"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "platforms": {
            "items": {
              "maxLength": 64,
              "pattern": "^[a-z0-9]+(_[a-z0-9]+)*$",
              "type": "string"
            },
            "title": "Platforms",
            "type": "array"
          },
          "type": {
            "maxLength": 64,
            "pattern": "^[a-z0-9]+(_[a-z0-9]+)*$",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "name",
          "type"
        ],
        "title": "Product",
        "type": "object"
      },
      "ProductProfile": {
        "additionalProperties": false,
        "properties": {
          "audience": {
            "$ref": "#/components/schemas/Audience"
          },
          "brand": {
            "$ref": "#/components/schemas/Brand"
          },
          "business_model": {
            "$ref": "#/components/schemas/BusinessModel"
          },
          "features": {
            "items": {
              "$ref": "#/components/schemas/Feature"
            },
            "title": "Features",
            "type": "array"
          },
          "measurement": {
            "$ref": "#/components/schemas/Measurement"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "product"
        ],
        "title": "ProductProfile",
        "type": "object"
      },
      "ProfileStatus": {
        "enum": [
          "draft",
          "approved"
        ],
        "title": "ProfileStatus",
        "type": "string"
      },
      "ProfileVersionResponse": {
        "additionalProperties": false,
        "properties": {
          "approved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved At"
          },
          "approved_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved By"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "profile": {
            "$ref": "#/components/schemas/ProductProfile"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ProfileStatus"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "project_id",
          "version",
          "status",
          "profile",
          "created_at",
          "approved_by",
          "approved_at"
        ],
        "title": "ProfileVersionResponse",
        "type": "object"
      },
...
      "SaveProfileEditRequest": {
        "additionalProperties": false,
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/ProductProfile"
          }
        },
        "required": [
          "profile"
        ],
        "title": "SaveProfileEditRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "OpenMarketer API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/projects/{project_id}/analyses": {
      "post": {
        "description": "Start analysing the repository and return the run to poll. The work happens later.",
        "operationId": "startAnalysis",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Start",
        "tags": [
          "analyses"
        ]
      }
    },
    "/v1/projects/{project_id}/analyses/{run_id}": {
      "get": {
        "description": "The run as it is now: queued, running, succeeded or failed.",
        "operationId": "getAnalysis",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get",
        "tags": [
          "analyses"
        ]
      }
    },
    "/v1/projects/{project_id}/profile/approved": {
      "get": {
        "description": "The approved version with the highest version number.",
        "operationId": "getApprovedProfile",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileVersionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Approved",
        "tags": [
          "profiles"
        ]
      }
    },
    "/v1/projects/{project_id}/profile/draft": {
      "get": {
        "description": "The draft with the highest version number. It can be older than the approved profile.",
        "operationId": "getDraftProfile",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileVersionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Draft",
        "tags": [
          "profiles"
        ]
      }
    },
    "/v1/projects/{project_id}/profile/versions/{version}/approval": {
      "post": {
        "description": "Approve one version as the current user. The request has no body.",
        "operationId": "approveProfileVersion",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Version",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileVersionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Approve",
        "tags": [
          "profiles"
        ]
      }
    },
    "/v1/projects/{project_id}/profile/versions/{version}/edits": {
      "post": {
        "description": "Store an edit of ``version`` as a new draft version. The edited version is not changed.",
        "operationId": "saveProfileEdit",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Version",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveProfileEditRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileVersionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Save Edit",
        "tags": [
          "profiles"
        ]
      }
    }
  }
}